+++ src/Routemeister/MessageHandlerCreator.cs
@@ -1,0 +1,6 @@
+﻿using System;
+
+namespace Routemeister
+{
+    public delegate object MessageHandlerCreator(Type messageHandlerContainerType, MessageEnvelope envelope);
+}